refactoring move sub model packages into common model package
