+++ src/main/java/org/icgc/argo/workflow_management/graphql/MutationDataFetcher.java
@@ -22,9 +22,9 @@ import com.google.common.collect.ImmutableMap;
 import graphql.schema.DataFetcher;
 import graphql.schema.DataFetchingEnvironment;
 import lombok.val;
-import org.icgc.argo.workflow_management.controller.model.RunsRequest;
-import org.icgc.argo.workflow_management.controller.model.RunsResponse;
-import org.icgc.argo.workflow_management.graphql.model.GqlRunsRequest;
+import org.icgc.argo.workflow_management.model.wes.RunsRequest;
+import org.icgc.argo.workflow_management.model.wes.RunsResponse;
+import org.icgc.argo.workflow_management.model.graphql.GqlRunsRequest;
 import org.icgc.argo.workflow_management.service.WorkflowExecutionService;
 import org.icgc.argo.workflow_management.service.model.WESRunParams;
 import org.springframework.beans.factory.annotation.Autowired;
