+++ src/main/java/org/icgc/argo/workflow_management/service/WorkflowExecutionService.java
@@ -20,7 +20,7 @@ package org.icgc.argo.workflow_management.service;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
-import org.icgc.argo.workflow_management.controller.model.RunsResponse;
+import org.icgc.argo.workflow_management.model.wes.RunsResponse;
 import org.icgc.argo.workflow_management.service.model.WESRunParams;
 import org.springframework.security.access.prepost.PreAuthorize;
 import reactor.core.publisher.Mono;
@@ -33,6 +33,6 @@ public interface WorkflowExecutionService {
   Mono<RunsResponse> cancel(String runId);
 
   @Retention(RetentionPolicy.RUNTIME)
-  @PreAuthorize("@queryAndMutationScopeChecker.apply(authentication. #runId)")
+  @PreAuthorize("@queryAndMutationScopeChecker.apply(authentication)")
   @interface HasQueryAndMutationAccess {}
 }
